Change options 'delete' and 'download' to 'clean_up' and 'fetch_all'.
'Clean_up': move files that are not in the local folder to an Archive folder.
'delete' function is ambiguous and dangerous. It may accidentally remove the wrong files (e.g copies of the original pdfs).

'fetch_all': fetch all files in the target folder from reMarkable, no matter whether they are already in the local folder.

* The warning info from RMAPI is removed from the file list.